requirements.txt
schedule==1.2.1
speedtest-cli==2.1.3
